requirements.txt
# Core Flask
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5

# Database
SQLAlchemy==2.0.23
alembic==1.13.1
# psycopg2-binary==2.9.9  # Only needed for production PostgreSQL, not SQLite

# Security & Auth
Werkzeug==3.0.1
WTForms==3.1.1
email-validator==2.1.0

# Media handling
Pillow==10.1.0

# Markdown
markdown==3.5.1
bleach==6.1.0

# Utilities
python-dotenv==1.0.0
gunicorn==21.2.0

# Development
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
